backend/requirements.txt
langchain
langgraph
langsmith
langchain_community
typing
typing_extensions
ipykernel
bs4
chromadb[all]
requests
sentence-transformers
langchain_groq
langchain_core
ipywidgets
huggingface_hub
faiss-cpu
sentencepiece
pyaudio
speechRecognition
python-dotenv
sounddevice
alive_progress
pygame
edge-tts
nest_asyncio
twilio
pydub
deepgram-sdk